Have to seperate out Base samplers into different modules because of moving imports to top of file.
